+++ NEWS.md
@@ -1,16 +1,19 @@
 # Development version
 
+* `decarbonate_ph`: function to remove CO2 (H2CO3) from a water and determine the new pH (and division of ions)
 * `modify_water`: function to modify individual slots in a water that handles unit conversions
 
 ## Breaking changes
 
+* `chemdose_ph_once` and `chemdose_toc_once` functions are back, and only return relevant water slots as new columns
 * Corrosion index slots removed from water. `calculate_corrosion` now outputs a data frame with the corrosion indices as columns. 
 `calculate_corrosion_chain` removed because the base function has a numeric output.
 * `solvect_o3` and `solveresid_o3` now return zero instead of NaN when the input dose is zero
 * CaSO4 added as an input option to `chemdose_ph`
 * `applied_treatment` slot removed from water. Was not providing any benefit and added complexity. Should be the responsibility of the user to track.
 * `_toc` functions are DOC-based and previously assumed no particulate TOC removal. In reality, virtually all particulate 
-TOC is removed with other particulates, so models have been updated to return TOC = DOC.
+TOC is removed with other particulates, so models have been updated to return TOC = DOC
+* `h2co3` slot added to `water` class calculated based on {H} and total carbonate.
 
 # tidywater 0.8.2